--- src/main/java/hanium/where2go/domain/restaurant/dto/RestaurantEnrollResponseDto.java
@@ -1,15 +1,0 @@
-package hanium.where2go.domain.restaurant.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class RestaurantEnrollResponseDto {
-
-    private Long restaurantId;
-    private String name;
-
-}
