--- src/main/java/hanium/where2go/domain/restaurant/dto/RestaurantMenuBoardUpdateRequestDto.java
@@ -1,17 +1,0 @@
-package hanium.where2go.domain.restaurant.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RestaurantMenuBoardUpdateRequestDto {
-
-    private String image_url;
-
-}
