--- src/main/java/hanium/where2go/domain/restaurant/dto/RestaurantMenuBoardUpdateResponseDto.java
@@ -1,16 +1,0 @@
-package hanium.where2go.domain.restaurant.dto;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RestaurantMenuBoardUpdateResponseDto {
-
-    private Long menu_board_id;
-}
